move code from candid_parser to candid_helper
